refactor(reactive-streams): use compression-intent pattern for safe third-party extensibility
