+++ rollbar-reactive-streams/src/main/java/com/rollbar/reactivestreams/notifier/sender/AsyncSender.java
@@ -10,12 +10,9 @@ import com.rollbar.reactivestreams.Utils;
 import com.rollbar.reactivestreams.notifier.sender.http.AsyncHttpClient;
 import com.rollbar.reactivestreams.notifier.sender.http.AsyncHttpRequest;
 import com.rollbar.reactivestreams.notifier.sender.http.AsyncHttpResponse;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.LinkedHashMap;
-import java.util.zip.GZIPOutputStream;
 import org.reactivestreams.Publisher;
 
 /**
@@ -56,13 +53,8 @@ public class AsyncSender implements Sender {
 
     String json = jsonSerializer.toJson(payload);
 
-    AsyncHttpRequest request;
-    if (compressPayload) {
-      headers.put("Content-Encoding", "gzip");
-      request = AsyncHttpRequest.Builder.build(this.url, headers.entrySet(), compress(json));
-    } else {
-      request = AsyncHttpRequest.Builder.build(this.url, headers.entrySet(), json);
-    }
+    AsyncHttpRequest request =
+        AsyncHttpRequest.Builder.build(this.url, headers.entrySet(), json, compressPayload);
 
     return Utils.map(httpClient.send(request),
         new Utils.Converter<AsyncHttpResponse, Response>() {
@@ -72,18 +64,6 @@ public class AsyncSender implements Sender {
             return new Response.Builder().result(result).status(from.getStatusCode()).build();
           }
         });
-  }
-
-  private static byte[] compress(String json) {
-    try {
-      ByteArrayOutputStream baos = new ByteArrayOutputStream();
-      GZIPOutputStream gzip = new GZIPOutputStream(baos);
-      gzip.write(json.getBytes(SyncSender.UTF_8));
-      gzip.close();
-      return baos.toByteArray();
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to gzip-compress payload", e);
-    }
   }
 
   @Override
