+++ rollbar-reactive-streams/src/main/java/com/rollbar/reactivestreams/notifier/sender/http/AsyncHttpRequest.java
@@ -13,19 +13,19 @@ public interface AsyncHttpRequest {
 
   String getBody();
 
-  default byte[] getBodyBytes() {
-    return null;
+  default boolean isCompressionRequested() {
+    return false;
   }
 
   class Builder {
     public static AsyncHttpRequest build(String url, Set<Map.Entry<String, String>> headers,
                                          String reqBody) {
-      return new AsyncHttpRequestImpl(url, headers, reqBody);
+      return new AsyncHttpRequestImpl(url, headers, reqBody, false);
     }
 
     public static AsyncHttpRequest build(String url, Set<Map.Entry<String, String>> headers,
-                                         byte[] body) {
-      return new AsyncHttpRequestImpl(url, headers, body);
+                                         String reqBody, boolean compressionRequested) {
+      return new AsyncHttpRequestImpl(url, headers, reqBody, compressionRequested);
     }
   }
 }
