+++ rollbar-reactive-streams/src/main/java/com/rollbar/reactivestreams/notifier/sender/http/AsyncHttpRequestImpl.java
@@ -9,22 +9,14 @@ class AsyncHttpRequestImpl implements AsyncHttpRequest {
   private final String url;
   private final Iterable<Map.Entry<String, String>> headers;
   private final String body;
-  private final byte[] bodyBytes;
+  private final boolean compressionRequested;
 
   public AsyncHttpRequestImpl(String url, Iterable<Map.Entry<String, String>> headers,
-                              String body) {
+                              String body, boolean compressionRequested) {
     this.url = url;
     this.headers = headers;
     this.body = body;
-    this.bodyBytes = null;
-  }
-
-  public AsyncHttpRequestImpl(String url, Iterable<Map.Entry<String, String>> headers,
-                              byte[] bodyBytes) {
-    this.url = url;
-    this.headers = headers;
-    this.body = null;
-    this.bodyBytes = bodyBytes;
+    this.compressionRequested = compressionRequested;
   }
 
   @Override
@@ -43,7 +35,7 @@ class AsyncHttpRequestImpl implements AsyncHttpRequest {
   }
 
   @Override
-  public byte[] getBodyBytes() {
-    return bodyBytes;
+  public boolean isCompressionRequested() {
+    return compressionRequested;
   }
 }
